Define FilterTable using `@class` instead of `@alias`
...and mark all fields as optional. This makes it feasible to define
individual function signatures for filter fields that match the correct
input type rather than the generic types Inline or Block. Remove a
section from the README which is invalidated by this change.

Also enable `false` to be returned optionally as a second filter
function return value (to leave the returned elements untouched).

See #2.