Overload `List:insert`
LuaLS only supports optional function parameters if they are trailing.
Therefore, the old definitino of `List:insert` was resulting in a
diagnostic if the method was used with only one argument (two if you
count `self`). Overload the function definition instead.

Note that argument type checking doesn't seem to work if the one
argument variant is used. I believe this is due to the use of the
`@generic` annotation and perhaps will be fixed one day.