Change filter args from `Filter` to `FilterTable`
Pandoc has a number of `walk` functions, which take a filter as an
argument and run that filter on the specified element. I had previously
assumed that one coul pass tables of filters, but this does not appear
to be the case. For these functions, replace the argument's `Filter`
type with `FilterTable`. See jgm/pandoc#10995.